+++ pom.xml
@@ -88,8 +88,14 @@
     		<artifactId>org.apache.commons.lang</artifactId>
     		<version>2.6</version>
 		</dependency>
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>RELEASE</version>
+            <scope>compile</scope>
+        </dependency>
 
-	</dependencies>
+    </dependencies>
 
 	<build>
 		<plugins>
